requirements.txt
fastapi
python-multipart
python-dotenv
pymupdf
langchain-core
langchain-groq
groq
pydantic
